Add files via upload
This folder contains Java programs that demonstrate the understanding of Fields, Constructors, Methods and Access Modifiers. The program simulates a pay-to-stay car park which has a ticket machine at the entrance and a cashier at the exit. When a driver enters the car park, they receive a ticket with the arrival time. When they leave, the driver gives the ticket to the cashier, the duration of the stay is calculated and from that, how much must be paid. 

The program uses similar classes as those found in the Java_Programs_01 folder with the addition of a TariffTable class that records parking tariffs for a pay-to-stay car park.