Added CowsAndBulls
Concerns the construction of a simple game called ‘Cows and Bulls’ that is
implemented using the NumberUtils class.
The game involves guessing a mystery 4-digit number. The number contains no repeat digits and no
zero.
With each guess, the number of correct digits – bulls – and the number of digits that would be
correct if they were in the right position – cows.
For example, say the mystery number is 8657 and the player guesses 4678, there is one bull (6), and
two cows (7, 8).