Polymorphism, Sub classes and Inheritance
Includes 3 programs, Vehicles, Menu and VectorGraphics that demonstrate the use of polymorphism, sub classes and inheritance in Java. 

The Vehicles program creates and outputs 3 simple objects where the classes for the second and third objects inherit from the class for the first object. The first object is a Vehicle with a number of passengers and color. The second is a Car with additional number of doors and the third object is a plane with a manufacturer and model number. 

The Menu program is a  program that stores some details of pizzas (menu item number, size, base, extra cheese, extra
garlic), curries (menu item number, size, curry type) and softdrinks (menu item number, size,
flavour, bottle or can) in a single array, with the options of listing all menu items or deleting a
particular menu item. The program continuously prompts the user to choose an option from a list and act on that
option, until the exit option is chosen. If a menu item is not found,
output "Not found" instead of "Done".

Finally, we have the VectorGraphics Program which manages a  set of graphical objects based on the commands in a specified file. This is a graphical equivalent of the famous edlin tool for text manipulation.