Binary Search Tree Implementation
Computer science experiment to compare the Binary Search Tree with a traditional unsorted array data structure using an application to manage daily vaccination numbers for a list of countries. A more elaborate application could offer additional functionality but we focus on just this feature for this experiment. The VaccineArrayApp reads in the attached file (vaccinations.csv) and stores the data items within a traditional array. There are 9919 data items. The VaccineBSTApp performs the same tasks but using a BST instead of an array. Instrumentation was used to count the number of comparison operations performed by the program.