HashTable Incomplete
The goal is to write a perfect has function for a set of textual data. The dataset provided is 36 student IDs stored in the names36.txt file.

The hash table implementation stores simple String data items, so can only track if an item is in the data structure or not. The hash function uses the weights (with values 0-4) to create a linear combination of the 9 characters in a student ID